Cast all times to long long and printf with %ll

+++ src/health.c
@@ -370,8 +370,8 @@ void *health_main(void *ptr) {
         if(unlikely(check_if_resumed_from_suspention())) {
             apply_hibernation_delay = 1;
 
-            info("Postponing alarm checks for %zd seconds, because it seems that the system was just resumed from suspension."
-            , hibernation_delay
+            info("Postponing alarm checks for %lld seconds, because it seems that the system was just resumed from suspension."
+            , (long long)hibernation_delay
             );
         }
 
@@ -384,8 +384,8 @@ void *health_main(void *ptr) {
 
             if(unlikely(apply_hibernation_delay)) {
 
-                info("Postponing health checks for %zd seconds, on host '%s'."
-                     , hibernation_delay
+                info("Postponing health checks for %lld seconds, on host '%s'."
+                     , (long long)hibernation_delay
                      , host->hostname
                 );
