+++ src/rrdpush.c
@@ -887,8 +887,8 @@ static int rrdpush_receive(int fd
     if(health_enabled != CONFIG_BOOLEAN_NO) {
         if(alarms_delay > 0) {
             host->health_delay_up_to = now_realtime_sec() + alarms_delay;
-            info("Postponing health checks for %zd seconds, on host '%s', because it was just connected."
-            , alarms_delay
+            info("Postponing health checks for %lld seconds, on host '%s', because it was just connected."
+            , (long long)alarms_delay
             , host->hostname
             );
         }
